requirements.txt
fastapi
uvicorn
pydantic
win10toast
websockets
python-dotenv
win11toast
winotify


